requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
fastf1>=3.3.0
httpx>=0.27.0
websockets>=12.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
xgboost
scikit-learn
lightgbm